Fix createEmployee function
Added an if statement to the function to fix the bug  where it returns an employee even when nameInput and jobTitleInput have no values. Somehow the UItextField reads it as an empty string rather that nil and passes the guard statement block allowing an employee to be returned. This creates an employee instance with empty name string  when enrol button is pressed.